requirements.txt
PyQt6
numpy
soundcard
comtypes
pycaw
pywin32
winsdk
Pillow
PyQt-Fluent-Widgets
